requirements.txt
requests==2.31.0
websocket-client==1.7.0
schedule==1.2.1
pyyaml==6.0.1
python-dateutil==2.8.2 